fix: do not split root module for async chunks
