+++ e2e/fixtures/code-splitting.complex/src/index.ts
@@ -4,6 +4,6 @@ import Button from 'antd/es/button';
 console.log(React, Button);
 
 import('./should-be-split').then((m) => console.log(m));
-import('./should-be-common').then((m) => console.log(m));
+import('./should-not-be-common').then((m) => console.log(m));
 import('./should-be-merged').then((m) => console.log(m));
 import('./other-dynamic').then((m) => console.log(m));
